requirements.txt
Django==4.2.7
django-environ==0.11.2
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
firebase-admin==7.1.0